Home

The home for TermLayouts.jl documentation.

Quickstart

Install TermLayouts:

pkg> add https://github.com/soumitradev/TermLayouts.jl.git

Run TermLayouts.jl by running the run() function in the REPL:

julia> using TermLayouts

julia> TermLayouts.run()
starting REPL...

Configuration

TermLayouts.jl comes with it's default config, which is:

[TermLayouts]
  [TermLayouts.panels.left]
  width = 70
  title = ""
  title_color = ""
  border_color = "red"

  [TermLayouts.panels.right]
  width = 30
  title = ""
  title_color = ""
  border_color = "blue"

TermLayouts' scoped configuration

If a TermLayouts config exists in the global scope, it will apply to all projects that run in the global scope. If a TermLayouts config file exists in the environment that has been activated, it use that config file. The config file needs to be in the working directory of that environment, and should be named LocalPreferences.toml

A TermLayouts config file is fairly simple. It can be described by the following two structures:

TermLayouts.TermLayoutsPreferences — Type

Describe the layout of Panels in TermLayouts

Fields

panel1: Left Panel

panel2: Right Panel

TermLayouts.PanelPrefs — Type

Describe the configuration of a Panel in TermLayouts

Fields

width: Percentage of the maximum width the panel takes up

title: Panel title text

title_color: Panel title text color. See here

border_color: Panel border color. See here

TermLayouts is limited to two panels as of now, due to limited types of panel implementations.

Width overflow

If your panel widths don't add up to 100 or less than 100, it will crop your gith panel by default

The config file can be described with the above two structs, but in reality looks something like:

[TermLayouts]
  [TermLayouts.panels.left]
  width = 80
  title = "REPL"
  title_color = "blue"
  border_color = "blue"

  [TermLayouts.panels.right]
  width = 20
  title = "Plots"
  title_color = "red"
  border_color = "red"

The TOML format is pretty self explanatory.