Internals

Documentation for TermLayouts.jl's internal types and functions.

TermLayouts.ColoredChar — Type

Character struct that is used to parse ANSI strings

TermLayouts.EditableString — Type

String structure to represent any string that can be manipulated using ANSI escape codes

TermLayouts.TermLayoutsState — Type

A struct to describe the state of TermLayouts

TermLayouts.Wrapper — Type

Workaround for https://github.com/julia-vscode/julia-vscode/issues/1940

TermLayouts.enterchar — Method

Add one character to the EditableString

TermLayouts.loadprefs — Method

Loads TermLayouts preferences from the environment

TermLayouts.read_key — Function

Read input from keyboard and handle relevant ANSI codes

TermLayouts.simplifyANSI — Method

Simplify an ANSI string into what ends up being displayed in the terminal

TermLayouts.start_eval_backend — Method

Start the backend that will evaluate our expressions