Internals

Documentation for TermLayouts.jl's internal types and functions.

TermLayouts.ColoredChar — Type

Character struct that is used to parse ANSI strings

TermLayouts.EditableString — Type

String structure to represent any string that can be manipulated using ANSI escape codes

TermLayouts.TermLayoutsState — Type

A struct to describe the state of TermLayouts

TermLayouts.Wrapper — Type

Workaround for https://github.com/julia-vscode/julia-vscode/issues/1940

TermLayouts.crop_backtrace — Method

Crop the error backtrace so that errors don't expose internal TermLayouts functions in errors

TermLayouts.display_repl_error — Method

Override the way errors are displayed

TermLayouts.enterchar — Method

Add one character to the EditableString

TermLayouts.evalrepl — Method

Evaluate expressions with custom backend, handle output and errors

TermLayouts.find_first_topelevel_scope — Method

Find the topmost non-TermLayouts scope in a backtrace

TermLayouts.get_main_mode — Function

Get the main mode of the REPL

TermLayouts.hook_repl — Method

Install REPL hooks to override how the backend evaluates expressions

TermLayouts.initializeState — Method

Initializes variables for TermLayouts

TermLayouts.loadprefs — Method

Loads TermLayouts preferences from the environment

TermLayouts.read_key — Function

Read input from keyboard and handle relevant ANSI codes

TermLayouts.run_with_backend — Method

Pass the expression into the channels

TermLayouts.simplifyANSI — Method

Simplify an ANSI string into what ends up being displayed in the terminal

TermLayouts.start_eval_backend — Method

Start the backend that will evaluate our expressions

TermLayouts.transform_backend — Method

Transform the backend to change of expressions are evaluated