Internals

Documentation for TermLayouts.jl's internal types and functions.

TermLayouts.ColoredChar — Type

Character struct that is used to parse ANSI strings.

TermLayouts.EditableString — Type

String structure to represent any string that can be manipulated using ANSI escape codes.

TermLayouts.TermLayoutsState — Type

A struct to describe the state of TermLayouts

TermLayouts.Wrapper — Type

Workaround for https://github.com/julia-vscode/julia-vscode/issues/1940

TermLayouts.carriage_return — Method

Enter a carriage return in the EditableString.

TermLayouts.cursor_down — Method

Move cursor down in the EditableString.

TermLayouts.cursor_horizontal_absolute — Method

Set the absolute horizontal position of the cursor in the EditableString.

TermLayouts.cursor_left — Method

Move cursor left in the EditableString.

TermLayouts.cursor_next_line — Method

Move the cursor to the next line in the EditableString.

TermLayouts.cursor_position — Method

Set the absolute position of the cursor in the EditableString.

TermLayouts.cursor_prev_line — Method

Move the cursor to the previous line in the EditableString.

TermLayouts.cursor_right — Method

Move cursor right in the EditableString.

TermLayouts.cursor_up — Method

Move cursor up in the EditableString.

TermLayouts.enterchar — Method

Add one character to the EditableString.

TermLayouts.entercolor — Method

Begin typing in a color into the EditableString.

TermLayouts.erase_in_display — Method

Delete parts of the display in the EditableString.

TermLayouts.erase_in_line — Method

Erase parts of the line the cursor is in inside the EditableString.

TermLayouts.loadprefs — Method

Loads TermLayouts preferences from the environment

TermLayouts.newline — Method

Insert a newline in the EditableString. Equivalent to printing a \n.

TermLayouts.parseANSI — Method

Parse ANSI text, and update the state of the EditableString.

TermLayouts.read_key — Function

Read input from keyboard and handle relevant ANSI codes

TermLayouts.run — Method

Activate TermLayouts, and spawn a new REPL session

TermLayouts.scroll_down — Method

Scroll down by n lines in the EditableString.

TermLayouts.scroll_up — Method

Scroll up by n lines in the EditableString.

TermLayouts.simplifyANSI — Method

Simplify an ANSI string into what ends up being displayed in the terminal.

TermLayouts.start_eval_backend — Method

Start the backend that will evaluate our expressions

TermLayouts.to_string — Method

Get the current state of the EditableString as a string.